+++ linear-regression/linreg.py
@@ -1,87 +1,70 @@
-#%%
-"""Simple example implementation of single-variable linear regression"""
-
-import pandas as pd
-import numpy as np
-from sklearn.model_selection import train_test_split
-import matplotlib.pyplot as plt
+import sys
 import time
+import pandas as pd
+from sklearn.model_selection import train_test_split
+from sklearn.preprocessing import StandardScaler
+from sklearn.preprocessing import PolynomialFeatures
+from sklearn.linear_model import LinearRegression
+from sklearn.metrics import mean_squared_error
 
-"""
-Linear Function
-h(x) = w * x + b
+start = time.process_time()
 
-h(x): linear prediction
-w: weight vector
-x: feature vector
-b: bias
+max_degree = 2
 
-Mean squared error (MSE) Function
-E = (1/2) * Σ(i=1)^n (yi - ^yi)^2
-"""
+# Read command-line arguments, establish default file path
+if len(sys.argv) == 1:
+    file_path = "../sample-data/5_features_2_degree_1000_samples.csv"
+elif len(sys.argv) == 2:
+    file_path = sys.argv[1]
+else:
+    print("Usage: python linreg.py <file_path>")
+    sys.exit(1)
 
+# Load in data
+data = pd.read_csv(file_path)
 
-def main():
-    print('Original function (without added noise):\n y = 2.50x + 4.0')
+# Separate feature and target data
+y = data['y']
+X = data.drop(columns=['y'])
 
-    data = pd.read_csv('../sample-data/linear_single_var_500.csv')
+# Transform polynomial features, if necessary
+if max_degree > 1:
+    poly = PolynomialFeatures(1)
+    poly.fit_transform(X)
 
-    # Feature variables returned as np array
-    X = data["X"].values
+# Split into training and test sets
+X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=115)
 
-    # Target variable returned as np array
-    y = data["y"].values
+# Fit a scaler to training data
+scaler = StandardScaler().fit(X_train)
 
-    # Split into train and test set
-    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=.5, random_state=115)
+# Scale training and test data
+X_train = scaler.transform(X_train)
+X_test = scaler.transform(X_test)
 
-    # Number of training samples, n
-    n = len(X_train)
+# Instantiate a new linear regression model
+linreg = LinearRegression()
 
-    # Initialize weight vector and bias at 0 (horizontal line)
-    w = 0
-    b = 0
+# Time model fitting
+train_start = time.process_time()
+linreg.fit(X_train, y_train)
+train_end = time.process_time()
 
-    # Maximum iterations (epochs)
-    max_iter = 1000
+# Make predictions
+y_pred = linreg.predict(X_test)
 
-    # Learning rate ("step" size)
-    L = 0.005
+# Print model parameters
+print("Weights:")
+i = 1
+for weight in linreg.coef_:
+    print(f"    Feature {i}: {weight:2f}")
+    i += 1
+print(f"Bias: {linreg.intercept_}\n")
 
-    # Start clock
-    start_time = time.process_time()
+# Calculate MSE and print model parameters
+mse = mean_squared_error(y_test, y_pred)
+print(f"Test MSE: {mse}\n")
 
-    # Iterate epochs, training model
-    for _ in range(max_iter):
-        # Calculate linear predictions (y-hat, or ^y)
-        y_pred = w * X_train + b
-
-        # Partial derivative of MSE, or E, with respect to w and b
-        dE_dw = (-2/n) * np.sum(X_train * (y_train - y_pred))
-        dE_db = (-2/n) * np.sum(y_train - y_pred)
-
-        # Update weights in opposite direction of highest slope, moving towards a local minima
-        w -= L * dE_dw
-        b -= L * dE_db
-
-    # End clock
-    end_time = time.process_time()
-
-    print(f'Resulting function:\n y = {round(w,2)}x + {round(b,2)}')
-    print(f'Training time: {round((end_time - start_time), 4)} seconds')
-
-    # Make test predictions on test set
-    y_pred = w * X_test + b
-
-    # Plot test data
-    plt.scatter(X_test, y_test)
-
-    # Plot linear predictions
-    plt.plot(X_test, y_pred, color='red')
-    plt.show()
-
-
-if __name__ == '__main__':
-    main()
-
-# %%
+print(f'Training Time: {(train_end - train_start):3f} seconds')
+end = time.process_time()
+print(f'Total CPU Time: {(end - start):3f} seconds')