+++ mlper/include/preprocessing.h
@@ -8,6 +8,6 @@ void poly_transform(struct Dataset *data, int degree);
 
 // Feature scaling
 void standardize(struct Dataset *data);
-void unstandardize(struct Dataset *data, double *feature_data, int num_entries);
+void unstandardize(struct Dataset *data, double *feature_data, int num_samples);
 
 #endif